requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
mysql-connector==2.2.9
requests==2.25.1
schedule==1.1.0
urllib3==1.26.4
xlwt==1.3.0
